refactor : BoardStatus -> GameResult
